Move ModalStackNavigators to a directory
